use usize for parts

+++ kernel/src/snapshot.rs
@@ -283,7 +283,7 @@ struct CheckpointMetadata {
     /// The number of actions that are stored in the checkpoint.
     pub(crate) size: i64,
     /// The number of fragments if the last checkpoint was written in multiple parts.
-    pub(crate) parts: Option<i32>,
+    pub(crate) parts: Option<usize>,
     /// The number of bytes of the checkpoint.
     pub(crate) size_in_bytes: Option<i64>,
     /// The number of AddFile actions in the checkpoint.
@@ -369,7 +369,7 @@ fn list_log_files_with_checkpoint(
         // we (may) need to drop commits that are before the _actual_ last checkpoint (that
         // is, commits between a stale _last_checkpoint and the _actual_ last checkpoint)
         commit_files.retain(|parsed_path| parsed_path.version > max_checkpoint_version);
-    } else if checkpoint_files.len() != checkpoint_metadata.parts.unwrap_or(1) as usize {
+    } else if checkpoint_files.len() != checkpoint_metadata.parts.unwrap_or(1) {
         return Err(Error::Generic(format!(
             "_last_checkpoint indicated that checkpoint should have {} parts, but it has {}",
             checkpoint_metadata.parts.unwrap_or(1),
